Move compile wait vibe-log helpers into compile_wait_log.go
compile_wait.go grew past the 500-line production-file limit after the
timeout/attach work. The VibeLog write and formatting helpers are a
separate concern from the wait loop, so relocate them without behavior
change.
