+++ cli/project-runner/internal/projectrunner/compile_wait.go
@@ -11,7 +11,6 @@ import (
 
 	clierrors "github.com/hatayama/unity-cli-loop/common/errors"
 	"github.com/hatayama/unity-cli-loop/common/tooldocs"
-	"github.com/hatayama/unity-cli-loop/common/vibelog"
 
 	"github.com/hatayama/unity-cli-loop/common/clicontract"
 	"github.com/hatayama/unity-cli-loop/common/clicore"
@@ -271,307 +270,4 @@ func shouldWaitForCompileStatus(err error, outcome unityipc.UnitySendOutcome) bo
 		return true
 	}
 	return outcome.RequestAccepted && clierrors.IsFinalResponseTimeoutError(err)
-}
-
-func writeCompileVibeLog(projectRoot string, buildEntry func() vibelog.CLIVibeLogEntry) {
-	if !vibelog.IsCLIVibeLogEnabled() {
-		return
-	}
-
-	writeCompileVibeLogEntry(projectRoot, buildEntry())
-}
-
-func writeCompileVibeLogEntry(projectRoot string, entry vibelog.CLIVibeLogEntry) {
-	_ = vibelog.WriteCLIVibeLog(projectRoot, entry)
-}
-
-func logCliDebugModeResolved(connection unityipc.Connection, command string) {
-	writeCompileVibeLog(connection.ProjectRoot, func() vibelog.CLIVibeLogEntry {
-		return vibelog.CLIVibeLogEntry{
-			Level:     "INFO",
-			Operation: "cli_debug_mode_resolved",
-			Message:   "Resolved CLI debug mode for the command.",
-			Context: map[string]any{
-				"command":          command,
-				"debug_enabled":    true,
-				"debug_source":     "env",
-				"project_identity": vibelog.ProjectIdentity(connection.ProjectRoot),
-				"cli_version":      clicontract.ProjectRunnerVersion(),
-			},
-		}
-	})
-}
-
-func logCompileRequestPrepared(
-	connection unityipc.Connection,
-	params map[string]any,
-	requestID string,
-	waitTimeout time.Duration,
-) {
-	writeCompileVibeLog(connection.ProjectRoot, func() vibelog.CLIVibeLogEntry {
-		reloadExternalSceneChanges := compileReloadExternalSceneChangesEnabled(params)
-		return vibelog.CLIVibeLogEntry{
-			Level:     "INFO",
-			Operation: "cli_compile_request_prepared",
-			Message:   "Prepared compile request parameters before dispatch.",
-			Context: map[string]any{
-				"command":                        clicore.CompileCommandName,
-				"request_id":                     requestID,
-				"wait_for_domain_reload":         true,
-				"force_recompile":                compileForceRecompileEnabled(params),
-				"reload_external_scene_changes":  reloadExternalSceneChanges,
-				"stop_on_external_scene_changes": !reloadExternalSceneChanges,
-				"project_identity":               vibelog.ProjectIdentity(connection.ProjectRoot),
-				"endpoint":                       connection.Endpoint.Address,
-				"timeout_ms":                     waitTimeout.Milliseconds(),
-				"poll_interval_ms":               compileWaitPollInterval.Milliseconds(),
-				"response_timeout_ms":            compileResponseTimeout.Milliseconds(),
-			},
-			CorrelationID: requestID,
-		}
-	})
-}
-
-func logCompileRequestSendResult(
-	connection unityipc.Connection,
-	requestID string,
-	outcome unityipc.UnitySendOutcome,
-	err error,
-	startedAt time.Time,
-) {
-	writeCompileVibeLog(connection.ProjectRoot, func() vibelog.CLIVibeLogEntry {
-		return vibelog.CLIVibeLogEntry{
-			Level:     compileRequestSendResultLogLevel(err),
-			Operation: "cli_compile_request_send_result",
-			Message:   "Recorded compile request dispatch outcome before status polling.",
-			Context: map[string]any{
-				"command":            clicore.CompileCommandName,
-				"request_id":         requestID,
-				"request_dispatched": outcome.RequestDispatched,
-				"request_accepted":   outcome.RequestAccepted,
-				"response_received":  err == nil && len(outcome.Result) > 0,
-				"response_timeout":   err != nil && clierrors.IsFinalResponseTimeoutError(err),
-				"transport_error":    clicore.ErrorMessage(err),
-				"elapsed_ms":         time.Since(startedAt).Milliseconds(),
-				"endpoint":           connection.Endpoint.Address,
-				"project_identity":   vibelog.ProjectIdentity(connection.ProjectRoot),
-				"outcome_total_ms":   outcome.Timing.Total.Milliseconds(),
-				"outcome_dial_ms":    outcome.Timing.Dial.Milliseconds(),
-				"outcome_write_ms":   outcome.Timing.Write.Milliseconds(),
-				"outcome_read_ms":    outcome.Timing.Read.Milliseconds(),
-				"outcome_decode_ms":  outcome.Timing.Decode.Milliseconds(),
-			},
-			CorrelationID: requestID,
-		}
-	})
-}
-
-func compileRequestSendResultLogLevel(err error) string {
-	if err == nil {
-		return "INFO"
-	}
-	return "WARNING"
-}
-
-func logCompileStatusPollStart(
-	options compileCompletionOptions,
-	startedAt time.Time,
-	deadline time.Time,
-) {
-	writeCompileVibeLog(options.connection.ProjectRoot, func() vibelog.CLIVibeLogEntry {
-		return vibelog.CLIVibeLogEntry{
-			Level:     "INFO",
-			Operation: "cli_compile_status_poll_start",
-			Message:   "Started polling Unity compile status.",
-			Context: compileWaitLogContext(options, startedAt, map[string]any{
-				"started_at":       startedAt.UTC().Format(time.RFC3339Nano),
-				"deadline_at":      deadline.UTC().Format(time.RFC3339Nano),
-				"project_identity": vibelog.ProjectIdentity(options.connection.ProjectRoot),
-			}),
-			CorrelationID: options.requestID,
-		}
-	})
-}
-
-func logCompileStatusPollObservedIfChanged(
-	options compileCompletionOptions,
-	startedAt time.Time,
-	attempt int,
-	status compileStatusResponse,
-	err error,
-	lastObservationKey *string,
-) {
-	if !vibelog.IsCLIVibeLogEnabled() {
-		return
-	}
-
-	observationKey := compileStatusObservationKey(status, err)
-	if observationKey == *lastObservationKey {
-		return
-	}
-
-	*lastObservationKey = observationKey
-	writeCompileVibeLogEntry(options.connection.ProjectRoot, vibelog.CLIVibeLogEntry{
-		Level:     compileStatusPollObservedLogLevel(err),
-		Operation: "cli_compile_status_poll_observed",
-		Message:   "Observed Unity compile status while polling.",
-		Context: compileWaitLogContext(options, startedAt, map[string]any{
-			"attempt":                      attempt,
-			"ready":                        status.Ready,
-			"has_result":                   status.HasResult,
-			"is_compiling":                 status.IsCompiling,
-			"is_updating":                  status.IsUpdating,
-			"is_domain_reload_in_progress": status.IsDomainReloadInProgress,
-			"message":                      status.Message,
-			"transport_error":              clicore.ErrorMessage(err),
-		}),
-		CorrelationID: options.requestID,
-	})
-}
-
-func compileStatusPollObservedLogLevel(err error) string {
-	if err == nil {
-		return "INFO"
-	}
-	return "WARNING"
-}
-
-func compileStatusObservationKey(status compileStatusResponse, err error) string {
-	return fmt.Sprintf(
-		"%t|%t|%t|%t|%t|%s|%s",
-		status.Ready,
-		status.HasResult,
-		status.IsCompiling,
-		status.IsUpdating,
-		status.IsDomainReloadInProgress,
-		status.Message,
-		clicore.ErrorMessage(err),
-	)
-}
-
-func logCompileStatusPollComplete(
-	options compileCompletionOptions,
-	startedAt time.Time,
-	attempts int,
-	status compileStatusResponse,
-) {
-	writeCompileVibeLog(options.connection.ProjectRoot, func() vibelog.CLIVibeLogEntry {
-		summary := compileResultLogSummary(status.Result)
-		return vibelog.CLIVibeLogEntry{
-			Level:     "INFO",
-			Operation: "cli_compile_status_poll_complete",
-			Message:   "Unity compile status polling returned the stored result.",
-			Context: compileWaitLogContext(options, startedAt, map[string]any{
-				"poll_attempts": attempts,
-				"success":       summary.success,
-				"error_count":   summary.errorCount,
-				"warning_count": summary.warningCount,
-			}),
-			CorrelationID: options.requestID,
-		}
-	})
-}
-
-func logCompileWaitTimedOut(
-	options compileCompletionOptions,
-	startedAt time.Time,
-	attempts int,
-	lastStatus compileStatusResponse,
-	lastErr error,
-) {
-	writeCompileVibeLog(options.connection.ProjectRoot, func() vibelog.CLIVibeLogEntry {
-		return vibelog.CLIVibeLogEntry{
-			Level:     "WARNING",
-			Operation: "cli_compile_status_poll_timeout",
-			Message:   "Timed out while polling Unity compile status.",
-			Context: compileWaitLogContext(options, startedAt, map[string]any{
-				"poll_attempts":        attempts,
-				"last_status":          compileStatusLogContext(lastStatus),
-				"last_transport_error": clicore.ErrorMessage(lastErr),
-				"project_identity":     vibelog.ProjectIdentity(options.connection.ProjectRoot),
-			}),
-			CorrelationID: options.requestID,
-		}
-	})
-}
-
-func logCompileWaitCancelled(
-	options compileCompletionOptions,
-	startedAt time.Time,
-	attempts int,
-	lastStatus compileStatusResponse,
-	lastErr error,
-	cancelErr error,
-) {
-	writeCompileVibeLog(options.connection.ProjectRoot, func() vibelog.CLIVibeLogEntry {
-		return vibelog.CLIVibeLogEntry{
-			Level:     "WARNING",
-			Operation: "cli_compile_status_poll_cancelled",
-			Message:   "Compile status polling was cancelled.",
-			Context: compileWaitLogContext(options, startedAt, map[string]any{
-				"poll_attempts":        attempts,
-				"last_status":          compileStatusLogContext(lastStatus),
-				"last_transport_error": clicore.ErrorMessage(lastErr),
-				"cancel_error":         clicore.ErrorMessage(cancelErr),
-			}),
-			CorrelationID: options.requestID,
-		}
-	})
-}
-
-func compileWaitLogContext(
-	options compileCompletionOptions,
-	startedAt time.Time,
-	extra map[string]any,
-) map[string]any {
-	context := map[string]any{
-		"command":          clicore.CompileCommandName,
-		"request_id":       options.requestID,
-		"force_recompile":  options.forceRecompile,
-		"endpoint":         options.connection.Endpoint.Address,
-		"timeout_ms":       options.timeout.Milliseconds(),
-		"poll_interval_ms": options.pollInterval.Milliseconds(),
-		"elapsed_ms":       time.Since(startedAt).Milliseconds(),
-	}
-	for key, value := range extra {
-		context[key] = value
-	}
-	return context
-}
-
-type compileResultSummary struct {
-	success      any
-	errorCount   any
-	warningCount any
-}
-
-func compileResultLogSummary(result json.RawMessage) compileResultSummary {
-	var payload map[string]any
-	if err := json.Unmarshal(result, &payload); err != nil {
-		return compileResultSummary{}
-	}
-	return compileResultSummary{
-		success:      firstPresentJSONValue(payload, "Success", "success"),
-		errorCount:   firstPresentJSONValue(payload, "ErrorCount", "errorCount"),
-		warningCount: firstPresentJSONValue(payload, "WarningCount", "warningCount"),
-	}
-}
-
-func firstPresentJSONValue(payload map[string]any, primaryKey string, legacyKey string) any {
-	value, ok := payload[primaryKey]
-	if ok {
-		return value
-	}
-	return payload[legacyKey]
-}
-
-func compileStatusLogContext(status compileStatusResponse) map[string]any {
-	return map[string]any{
-		"ready":                        status.Ready,
-		"has_result":                   status.HasResult,
-		"is_compiling":                 status.IsCompiling,
-		"is_updating":                  status.IsUpdating,
-		"is_domain_reload_in_progress": status.IsDomainReloadInProgress,
-		"message":                      status.Message,
-	}
 }
